Split out deleteSite into wizard steps
